Attempt to combine DatePickerStateful and DateTimeInput (not working)
